+++ R/emm.R
@@ -60,6 +60,9 @@ emm <- function(model, ci.lvl = .95, type = c("fe", "re"), typical = "mean", ...
       ppd = FALSE,
       terms = sjstats::pred_vars(model),
       typical,
+      vcov.fun = NULL,
+      vcov.type = NULL,
+      vcov.args = NULL,
       ...
     )